requirements.txt
google-generativeai
streamlit
vertexai
sentence-transformers
requests
python-dotenv
oath2client